refactor: new send form file structure
