+++ src/app/pages/send/choose-crypto-asset/components/crypto-currency-asset-icon.tsx
