+++ src/app/pages/send/choose-crypto-asset/components/fungible-token-asset-item.tsx
