+++ src/app/pages/send/send-crypto-asset-form/components/send-crypto-asset.layout.tsx
@@ -3,7 +3,7 @@ import { Flex } from '@stacks/ui';
 interface SendCryptoAssetLayoutProps {
   children: JSX.Element;
 }
-export function SendCryptoAssetLayout({ children }: SendCryptoAssetLayoutProps) {
+export function ChooseCryptoAssetLayout({ children }: SendCryptoAssetLayoutProps) {
   return (
     <Flex alignItems="center" flexDirection="column" justifyContent="center" mt="48px" width="100%">
       {children}
